Moved helper functions to top of test files.
